small refactor and logic cleanup
